+++ tests/jest/plugins/search/plugin.search.test.js
@@ -1,7 +1,7 @@
 import BookReader from '@/src/BookReader.js';
-import '@/src/plugins/plugin.mobile_nav.js';
-import { marshallSearchResults } from '@/src/plugins/search/plugin.search.js';
+import '@/src/plugins/search/plugin.search.js';
 import { deepCopy } from '../../utils.js';
+import { DUMMY_RESULTS } from './utils.js';
 
 jest.mock('@/src/plugins/search/view.js');
 
@@ -12,28 +12,6 @@ const triggeredEvents = () => {
     if (typeof params[0] === 'string') { return params[0]; }
     return params[0].type;
   });
-};
-
-const DUMMY_RESULTS = {
-  ia: "adventuresofoli00dick",
-  q: "child",
-  indexed: true,
-  page_count: 644,
-  body_length: 666,
-  leaf0_missing: false,
-  matches: [{
-    text: 'For a long; time after it was ushered into this world of sorrow and trouble, by the parish surgeon, it remained a matter of considerable doubt wliether the {{{child}}} Avould survi^ e to bear any name at all; in which case it is somewhat more than probable that these memoirs would never have appeared; or, if they had, that being comprised within a couple of pages, they would have possessed the inestimable meiit of being the most concise and faithful specimen of biography, extant in the literature of any age or country.',
-    par: [{
-      boxes: [{r: 1221, b: 2121, t: 2075, page: 37, l: 1107}],
-      b: 2535,
-      t: 1942,
-      page_width: 1790,
-      r: 1598,
-      l: 50,
-      page_height: 2940,
-      page: 37
-    }]
-  }]
 };
 
 beforeEach(() => {
@@ -163,20 +141,5 @@ describe('Plugin: Search', () => {
     br.init();
     await br.search('foo');
     expect(triggeredEvents()).toContain(`${namespace}SearchCallbackEmpty`);
-  });
-});
-
-describe('marshallSearchResults', () => {
-  test('Adds match index', () => {
-    const results = deepCopy(DUMMY_RESULTS);
-    marshallSearchResults(results, x => x.toString());
-    expect(results.matches[0]).toHaveProperty('matchIndex', 0);
-    expect(results.matches[0].par[0].boxes[0]).toHaveProperty('matchIndex', 0);
-  });
-
-  test('Adds display page number', () => {
-    const results = deepCopy(DUMMY_RESULTS);
-    marshallSearchResults(results, x => `n${x}`);
-    expect(results.matches[0]).toHaveProperty('displayPageNumber', 'n37');
   });
 });
