+++ tests/jest/plugins/search/plugin.search.view.test.js
@@ -2,6 +2,7 @@
 import BookReader from '@/src/BookReader.js';
 import '@/src/plugins/plugin.mobile_nav.js';
 import '@/src/plugins/search/plugin.search.js';
+import { marshallSearchResults } from '@/src/plugins/search/utils.js';
 import '@/src/plugins/search/view.js';
 
 let br;
@@ -27,6 +28,8 @@ const results = {
     }]
   }]
 };
+
+marshallSearchResults(results, () => '', '{{{', '}}}');
 const resultWithScript = {
   ia: "adventuresofoli00dick",
   q: "child",
@@ -48,6 +51,8 @@ const resultWithScript = {
     }]
   }]
 };
+
+marshallSearchResults(resultWithScript, () => '', '{{{', '}}}');
 beforeEach(() => {
   $.ajax = jest.fn().mockImplementation(() => {
     // return from:
